Fixed No-Record Fallback
The client expects a response object “bad_team”, which is used to
discriminate between a non-responsive database and a simple lack of
data. This was not functioning properly before. Now, trying to access a
teams data that we don’t have will result in a response of {bad_team:
(teamNumber)}. If the client receives it, the client will search The
Blue Alliances’ database.