Write functionality - significant progress.
It’s not done yet, but it is now possible to write to the database.
Still working out bugs as I find them, but things look all right as of
now. It is only possible to write to the scout:pit section right now,
but the rest is smooth sailing. I need to start thinking about security
and user authentication. Ugh.