Allow Asynchronous Component Rendering
There was trouble before with the way that the style was defined. It is
now properly distributed between the constructor and the render method,
so that any changes to the component’s props after it’s rendering will
be reflected.

This is particularly important for the rendering of analytic data that
will be gathered asynchronously with Fetch.