Add functionality for options from competition landing
Button options for the different input pages will appear, and selecting
one will lock that team as the selected option - the team will not be
accidentally switched.

Additionally, PitScout now shows team names as well as numbers. If
there is no record, The Blue Alliance is called. Like before, this is
not ideal, but it’s better than nothing.

Finally, a little bit of housekeeping was done. Removed some leftover
debugging things and changed the competition to Reading, including the
official teams.