made some more changes to the server address system
So…
This is what I would call a creative solution. I wanted to be able to
transfer data to the node server over https, but wasn’t able to get a
proper cert for it. My solution was a PhP file which cURLs the server.
Of course, this means two servers have to be running - one to host the
files (apache in my case) and another for the actual api. Would it have
been better to make the whole thing in PhP? Maybe. I wasn’t really
thinking this far ahead when I made it though. This solution works
fine, with the one exception being that requests need to be made to
[host]/[url] now, instead of [host]:8081. I think that this is actually
more logical anyways, though.